backend/requirement.txt
Flask==2.3.2
Flask-SocketIO==5.3.2
Flask-Cors==3.0.10
flask-sqlalchemy
eventlet==0.33.3
tensorflow==2.17.0
keras==3.5.0
numpy==1.24.3
pillow==9.5.0
opencv-python==4.8.0.74
pytesseract==0.3.10
sqlalchemy==2.0.0
python-multipart==0.0.6
fastapi==0.100.0
uvicorn==0.23.2
olefile

